+++ response/common_response.go
@@ -1,0 +1,71 @@
+package commonresp
+
+import (
+	"net/http"
+
+	commonerr "github.com/kodiakdev/go-common-lib/err"
+
+	"github.com/emicklei/go-restful"
+	"github.com/sirupsen/logrus"
+)
+
+// RequestResponse holds all value of request and response
+type RequestResponse struct {
+	Req          *restful.Request
+	Resp         *restful.Response
+	HTTPStatus   int
+	ResponseBody interface{}
+}
+
+//ServiceErrorResponse response commonerr for non 2xx
+type ServiceErrorResponse struct {
+	Code        int    `json:"code"`
+	Explanation string `json:"explanation"`
+}
+
+//write perform write the response as JSON
+func write(comm *RequestResponse) {
+	err := comm.Resp.WriteHeaderAndJson(
+		comm.HTTPStatus,
+		comm.ResponseBody,
+		restful.MIME_JSON,
+	)
+	if err != nil {
+		logrus.Warnf("Unable to write response. Error was %v", err)
+	}
+}
+
+func respondRequestParsingFail(err error, req *restful.Request, resp *restful.Response) {
+	logrus.Warnf("Failed to read entity. Error: %v", err)
+	errorResponseBody := ServiceErrorResponse{
+		Code:        commonerr.FailedParseRequestBodyCode,
+		Explanation: commonerr.FailedParseRequestBodyExplanation,
+	}
+	respond(errorResponseBody, http.StatusBadRequest, req, resp)
+}
+
+func respondDatabaseError(err error, req *restful.Request, resp *restful.Response) {
+	errorResponseBody := ServiceErrorResponse{
+		Code:        commonerr.DatabaseErrorCode,
+		Explanation: commonerr.DatabaseErrorExplanation,
+	}
+	respond(errorResponseBody, http.StatusInternalServerError, req, resp)
+}
+
+func respondUnknownError(err error, req *restful.Request, resp *restful.Response) {
+	logrus.Errorf("Error occured with unknown reason. Error: %v", err)
+	errorResponseBody := ServiceErrorResponse{
+		Code:        commonerr.UnknownErrorCode,
+		Explanation: commonerr.UnknownErrorExplanation,
+	}
+	respond(errorResponseBody, http.StatusInternalServerError, req, resp)
+}
+
+func respond(body interface{}, httpStatus int, req *restful.Request, resp *restful.Response) {
+	write(&RequestResponse{
+		Req:          req,
+		Resp:         resp,
+		HTTPStatus:   httpStatus,
+		ResponseBody: body,
+	})
+}