+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.0.1-alpha.3
+### Added
+- Common repo db operation: Find, FindOne, Insert, Update
+- Common response across services such as UnknownError or DBProblem response
+- Common error code and explanation across services
+- Swagger html error message generator
+
 ## 0.0.1-alpha.2
 ### Fixed
 - Module declaration in go.mod